Pivot to HuggingFace Diffusers as reference implementation
Major refactoring from mlx-video (third-party port) to HuggingFace Diffusers
as the canonical reference for LTX-2 text-to-video generation on Apple Silicon.

Key changes:
- Per-component weight loading from Diffusers format (transformer/, vae/,
  connectors/, text_encoder/) instead of single unified file
- 3-phase memory management: text encoding → unload Gemma → denoising →
  unload transformer → VAE decode (never more than 1 large model in memory)
- MemoryOptimizationConfig with presets (disabled/light/moderate/aggressive)
- Float32 latents throughout pipeline, cast to bf16 only for transformer input
- Attention mask propagation to transformer (fixes text conditioning)
- Cross-attention pre-norm removal (matching Diffusers architecture)
- Fixed VAE weight mapping (resnets → res_blocks renaming)
- Fixed EOS detection for Gemma 3 prompt enhancement (stop on <end_of_turn>)
- Standalone text encoding API: pipeline.encodeText() + CLI encode command
- Dev model (40 steps, CFG 4.0) and distilled model (8 steps, no CFG)
- Removed FP8 quantized variant (distilledFP8)
- SpatialUpscaler for two-stage generation pipeline
- RoPE with double-precision (float32) computation

Validated end-to-end: 512x512 at 25 and 97 frames with native text encoding
matching HuggingFace reference within bf16 precision (0.001-0.003 error).

+++ Sources/LTXVideo/Models/Transformer/LTXAttention.swift
@@ -24,8 +24,11 @@ public class RMSNorm: Module, UnaryLayer {
 }
 
 /// Apply RMS normalization without learnable weight
+///
+/// Matches Python: `mx.fast.rms_norm(x, mx.ones((x.shape[-1],), dtype=x.dtype), eps)`
+/// Weight dtype must match input dtype to avoid float32 promotion.
 public func rmsNorm(_ x: MLXArray, eps: Float = 1e-6) -> MLXArray {
-    let weight = MLXArray.ones([x.dim(-1)])
+    let weight = MLXArray.ones([x.dim(-1)]).asType(x.dtype)
     return MLXFast.rmsNorm(x, weight: weight, eps: eps)
 }
